feat: checklist in expansion tile
