+++ lib/features/tasks/ui/checkbox_item_widget.dart
@@ -1,6 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:lotti/features/tasks/state/checklist_item_controller.dart';
 
 // ignore: avoid_positional_boolean_parameters
 typedef BoolCallback = void Function(bool);
@@ -53,57 +51,6 @@ class _CheckboxItemWidgetState extends State<CheckboxItemWidget> {
         });
         widget.onChanged(_isChecked);
       },
-    );
-  }
-}
-
-class CheckboxItemWrapper extends ConsumerWidget {
-  const CheckboxItemWrapper(
-    this.itemId, {
-    super.key,
-  });
-
-  final String itemId;
-
-  @override
-  Widget build(
-    BuildContext context,
-    WidgetRef ref,
-  ) {
-    final provider = checklistItemControllerProvider(id: itemId);
-    final item = ref.watch(provider);
-
-    return item.map(
-      data: (data) {
-        final item = data.value;
-        if (item == null) {
-          return const SizedBox();
-        }
-        return CheckboxItemWidget(
-          title: item.data.title,
-          isChecked: item.data.isChecked,
-          onChanged: (checked) => ref.read(provider.notifier).toggleChecked(),
-        );
-      },
-      error: ErrorWidget.new,
-      loading: (_) => const CircularProgressIndicator(),
-    );
-  }
-}
-
-class CheckboxItemsList extends StatelessWidget {
-  const CheckboxItemsList({
-    required this.itemIds,
-    super.key,
-  });
-
-  final List<String> itemIds;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: itemIds.map(CheckboxItemWrapper.new).toList(),
     );
   }
 }
